requirements.txt
asgiref==3.3.1
Django==3.0.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
Markdown==3.3.4
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
